add core and plugin

+++ core/src/main/java/src/main/java/eu/ensg/hiba/fibonacci/core/SuiteFibonnaci.java
@@ -2,7 +2,7 @@ package src.main.java.eu.ensg.hiba.fibonacci.core;
 
 public class SuiteFibonnaci
 {
-    public static int fibo(int n)
+    public static long fibo(long n)
     {
         if (n < 2) return(n);
         return( fibo(n-2) + fibo(n-1) );
